+++ routes/elevators.ts
@@ -1,6 +1,6 @@
 import httpStatus from "http-status";
 import Elevator from "../models/Elevator";
-import { Router as elevatorRoute } from "./route";
+import { Router as elevatorRoute } from "../route";
 
 // Create Elevator
 elevatorRoute.post("/", async (req, res) => {
