+++ routes/rooms.ts
@@ -2,7 +2,7 @@ import httpStatus from "http-status";
 import { authenticateJWT } from "../jwtAuth";
 import Room from "../models/Room";
 import User from "../models/User";
-import { Router as roomRoute } from "./route";
+import { Router as roomRoute } from "../route";
 
 roomRoute.post("/", authenticateJWT, async (req, res) => {
   try {
