--- routes/route.ts
@@ -1,3 +1,0 @@
-import { Router as ExpressRouter } from "express";
-
-export const Router = ExpressRouter();
