+++ routes/schedule.ts
@@ -6,7 +6,7 @@ import TimeTable from "../models/Timetable";
 import Subject from "../models/Subject";
 import Class from "../models/Class";
 import Room from "../models/Room";
-import { Router as scheduleRoute } from "./route";
+import { Router as scheduleRoute } from "../route";
 
 export const createScheduleFromTimetable = async (
   req: Request,
