+++ routes/subjects.ts
@@ -1,7 +1,7 @@
 import httpStatus from "http-status";
 import Post from "../models/Post";
 import Subject from "../models/Subject";
-import { Router as subjectRoute } from "./route";
+import { Router as subjectRoute } from "../route";
 
 // Subject作成
 subjectRoute.post("/", async (req, res) => {
