+++ server.ts
@@ -1,25 +1,11 @@
 import dotenv from "dotenv";
+import mongoose from "mongoose";
 import path from "path";
+import { app } from "./route";
 
 dotenv.config({ path: path.resolve(__dirname, "../.env") });
 
-import cors from "cors";
-import express from "express";
-const app = express();
-
-import authRoute from "./routes/auth";
-import classRoute from "./routes/classes";
-import eventRoute from "./routes/event";
-import postRoute from "./routes/posts";
-import roomRoute from "./routes/rooms";
-import scheduleRoute from "./routes/schedule";
-import subjectRoute from "./routes/subjects";
-import teacherRoute from "./routes/teachers";
-import timetableRoute from "./routes/timetable";
-import userRoute from "./routes/users";
-
 const PORT = 4000;
-import mongoose from "mongoose";
 
 //データベース接続
 mongoose
@@ -31,18 +17,5 @@ mongoose
     console.log(err);
   });
 
-//ミドルウェア
-app.use(express.json());
-app.use(cors());
-app.use("/v1/users", userRoute);
-app.use("/v1/auth", authRoute);
-app.use("/v1/posts", postRoute);
-app.use("/v1/subjects", subjectRoute);
-app.use("/v1/rooms", roomRoute);
-app.use("/v1/teachers", teacherRoute);
-app.use("/v1/classes", classRoute);
-app.use("/v1/timetable", timetableRoute);
-app.use("/v1/schedule", scheduleRoute);
-app.use("/v1/event", eventRoute);
-
+// サーバースタート
 app.listen(PORT, () => console.log("サーバーが起動しました"));
